Fix undefined behavior pointer alignment in HandleGeneralPacket.
Use correct endian conversion function for uint16 (short).

Simplify ReadFIFOMonitorChannel with bitshift, use memcpy to fix alignment issue, change Current to uint16_t to match actual data size.

Add general register read/write mutex to more register read/mutate functions.

Remove unneeded Address and Channel from SetupFIFOMonitorChannel.

Enable -O2 optimization level by default.

Cleanup old header and rateword check code.

+++ sw_projects/common/saturndrivers.h
@@ -45,8 +45,7 @@ void SetupFIFOMonitorChannel(EDMAStreamSelect Channel, bool EnableInterrupt);
 //   Underflowed:       true if underflow has occurred. Cleared by read.
 //   Current:           number of locations occupied (in either FIFO type)
 //
-uint32_t ReadFIFOMonitorChannel(EDMAStreamSelect Channel, bool* Overflowed, bool* OverThreshold, bool* Underflowed, unsigned int* Current);
-
+uint32_t ReadFIFOMonitorChannel(EDMAStreamSelect Channel, bool* Overflowed, bool* OverThreshold, bool* Underflowed, uint16_t* Current);
 
 //
 // reset a stream FIFO
